+++ CHANGELOG.md
@@ -10,6 +10,7 @@
     - `BrowserPlugin` - Manages the browser automation framework, and basically acts as a factory for controllers.
 - Browser rotation with a maximum number of pages opened per browser.
 - Add emit persist state event to event manager
+- Add batched request addition in `RequestQueue`
 
 ## [0.0.4](../../releases/tag/v0.0.4) - 2024-05-30